# openapi.yaml
openapi: 3.1.0
info:
  title: NotebookLM OSS Clone API
  version: 0.1.0
  description: |
    Backend API for the NotebookLM OSS Clone. All routes are served by Next.js
    Route Handlers. The server proxies document ingestion, chat, and search to
    an OpenRAG instance; notebook/source/note metadata is managed locally.

    REQ-ID references map each endpoint to the requirements in requirements.md.

    ## Frontend-only requirements (no API endpoint)

    The following REQ-IDs are satisfied entirely in the browser and have no
    corresponding API endpoint:

    - REQ-005  Three-panel layout (CSS/component structure)
    - REQ-010  Source selection checkboxes (React state; effect carried via
               `selectedSourceIds` in POST /chat and POST /generate/* bodies)
    - REQ-011  Source list search filter (client-side string filter, no server call)
    - REQ-014  Chat history display (React state; conversation continuity via
               `chatId` round-trip in the chat endpoint)
    - REQ-015  Loading indicator while streaming (React state; signals are the
               first SSE token arriving and the `done` event closing the stream)
    - REQ-022  Empty state guidance (conditional rendering based on source count)
    - REQ-023  Error communication (frontend displays errors from API responses;
               the API contract is defined by the reusable error response schemas)
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/notebooks/{notebookId}:
    parameters:
      - $ref: "#/components/parameters/notebookId"

    get:
      operationId: getNotebook
      summary: Get a notebook
      description: "REQ-002 — Fetches a single notebook with its source count and note count."
      tags: [notebooks]
      responses:
        "200":
          description: Notebook detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      operationId: updateNotebook
      summary: Rename a notebook
      description: "REQ-003 — Updates the notebook name. Empty name is rejected."
      tags: [notebooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  example: "Advanced Physics"
      responses:
        "200":
          description: Updated notebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notebook"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteNotebook
      summary: Delete a notebook
      description: |
        REQ-004 — Deletes the notebook, all its sources (including removing them
        from OpenRAG), all its notes, and its OpenRAG knowledge filter.
      tags: [notebooks]
      responses:
        "204":
          description: Notebook deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Sources ────────────────────────────────────────────────────────────────
...
  /api/notebooks/{notebookId}/sources/text:
    parameters:
      - $ref: "#/components/parameters/notebookId"

    post:
      operationId: addTextSource
      summary: Add a plain-text source
      description: |
        REQ-006 — Accepts pasted text, ingests it to OpenRAG as a .txt file,
        and updates the notebook knowledge filter.
        Returns 409 if identical text content already exists as a source.
      tags: [sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                title:
                  type: string
                  description: "Optional. Defaults to first line of body, truncated to 60 chars."
                body:
                  type: string
                  minLength: 1
                  description: Plain text content.
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/notebooks/{notebookId}/sources/url:
    parameters:
      - $ref: "#/components/parameters/notebookId"

    post:
      operationId: addUrlSource
      summary: Add a URL source
      description: |
        REQ-008 — Fetches the page server-side, extracts readable text, ingests
        to OpenRAG as .txt, and updates the notebook knowledge filter.
        Returns 409 if the same URL already exists as a source.
        Returns 422 if the URL is unreachable or yields no readable text.
      tags: [sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  example: "https://en.wikipedia.org/wiki/Retrieval-augmented_generation"
                title:
                  type: string
                  description: "Optional. Defaults to the page <title> element."
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  parameters:
    notebookId:
      name: notebookId
      in: path
      required: true
      schema:
        type: string
      description: UUID of the notebook.
...
  schemas:
    NotebookSummary:
      type: object
      required: [id, name, sourceCount, noteCount, createdAt]
      properties:
        id:
          type: string
          example: "nb_abc123"
        name:
          type: string
          example: "Physics Notes"
        sourceCount:
          type: integer
          minimum: 0
        noteCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    Notebook:
      allOf:
        - $ref: "#/components/schemas/NotebookSummary"
        - type: object
          required: [openragFilterId]
          properties:
            openragFilterId:
              type: string
              description: ID of the associated OpenRAG knowledge filter.
              example: "filter_xyz789"
    Source:
      type: object
      required: [id, notebookId, title, type, openragFilename, createdAt]
      properties:
        id:
          type: string
          example: "src_def456"
        notebookId:
          type: string
          example: "nb_abc123"
        title:
          type: string
          example: "Lecture 3 — Thermodynamics"
        type:
          type: string
          enum: [text, pdf, url]
        url:
          type: ["string", "null"]
          format: uri
          description: "Populated for type=url only."
        openragFilename:
          type: string
          description: Filename key used in OpenRAG for this document.
          example: "nb_abc123-src_def456.pdf"
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "Notebook not found."
        code:
          type: ["string", "null"]
          description: Optional machine-readable error code.
          example: "NOTEBOOK_NOT_FOUND"

  # ── Reusable Responses ─────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: "400 — Invalid input (missing required field, empty string, wrong type)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "name must not be empty."
            code: "VALIDATION_ERROR"
    NotFound:
      description: "404 — The requested resource does not exist."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Notebook not found."
            code: "NOT_FOUND"
    Conflict:
      description: "409 — A duplicate resource already exists."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "A source with this filename already exists in the notebook."
            code: "DUPLICATE_SOURCE"
    UnprocessableEntity:
      description: |
        422 — The request was well-formed but could not be processed.
        Examples: PDF is password-protected, URL returned no readable text,
        AI could not extract tabular data, fewer than two concepts identified.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Could not extract readable text from the provided URL."
            code: "UNPROCESSABLE"
...
    InternalServerError:
      description: "500 — Unexpected server error."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "An unexpected error occurred."
            code: "INTERNAL_ERROR"